Put into action SQL specification and mapper in DAL
- Make an interface `Mapper` and set apart its implementation to standalone class
- Implement `SqlSpecification` interface with `String toSqlQuery()` method
