+++ client/src/app/search-form/search-form.component.html
@@ -6,6 +6,7 @@
       value="faire"
       [ngModelOptions]="{standalone: true}"
     />
+    <small>{{ errorMsg }}</small>
   </mat-form-field>
   <button
     mat-stroked-button
