+++ client/src/app/services/api.service.ts
@@ -1,7 +1,8 @@
 import {Injectable} from '@angular/core';
-import {HttpClient} from "@angular/common/http";
+import {HttpClient, HttpErrorResponse} from "@angular/common/http";
 import {Observable} from "rxjs";
 import {BaseVerb} from "../models/baseVerb";
+import {catchError} from "rxjs/operators";
 
 @Injectable()
 export class ApiService {
@@ -13,11 +14,12 @@ export class ApiService {
 
   public getOneVerb(verbToSearch: string = "aller"): Observable<BaseVerb> {
     return this.http
-      .get<BaseVerb>("http://localhost:8080/api/find/" + verbToSearch);
+      .get<BaseVerb>("http://localhost:8080/api/find/" + verbToSearch)
+      .pipe(catchError(this.handleError));
   }
 
-  private handleError(error: Response) {
-    return Observable.throw(error || 'Server Error');
+  private handleError(error: HttpErrorResponse) {
+    return Observable.throw(error.message || 'Server Error');
   }
 
 }
