+++ server/src/main/java/dev/aubique/conj/mappers/Mapper.java
@@ -1,0 +1,8 @@
+package dev.aubique.conj.mappers;
+
+import java.sql.SQLException;
+
+@FunctionalInterface
+public interface Mapper<From, To> {
+    To map(From from) throws SQLException;
+}
