+++ server/src/main/java/dev/aubique/conj/model/Verb.java
@@ -1,4 +1,4 @@
-package dev.aubique.conj.models;
+package dev.aubique.conj.model;
 
 import java.util.List;
