+++ server/src/main/webapp/WEB-INF/web.xml
@@ -6,7 +6,7 @@
     <display-name>Archetype Created Web Application</display-name>
 
     <listener>
-        <listener-class>dev.aubique.conj.listeners.ServletContextListenerImpl</listener-class>
+        <listener-class>dev.aubique.conj.listener.ServletContextListenerImpl</listener-class>
     </listener>
 
     <servlet>
